+++ app/components/workout/index.ts
@@ -9,3 +9,4 @@ export * from './SessionFeedbackModal';
 export * from './AdHocWorkoutModal';
 export * from './SorenessCheckIn';
 export * from './EditableSetTable';
+export * from './SessionDetailModal';